refactor: progressRate를 Int형에서 Float형으로 수정

+++ android/shared/src/commonMain/kotlin/com/yeobaek/feature/home/model/CurrentlyReadingBookUiModel.kt
@@ -5,5 +5,5 @@ data class CurrentlyReadingBookUiModel(
     val title: String,
     val coverImageUrl: String,
     val authors: String,
-    val progressRate: Int,
+    val progressRate: Float,
 )
